Enhance fim_run_realtime function and add UTs.

+++ src/syscheckd/config.c
@@ -142,7 +142,7 @@ int Read_Syscheck_Config(const char *cfgfile)
     if ((syscheck.directories == NULL) && (syscheck.registry[0].entry == NULL)) {
         return (1);
     }
-    syscheck.max_fd_win_rt = getDefine_Int("syscheck", "max_fd_win_rt", 1, 1024);
+    syscheck.max_fd_win_rt = (unsigned int) getDefine_Int("syscheck", "max_fd_win_rt", 1, 1024);
 #endif
 
     return (0);
